Workable start/stop grid scrips for single node

+++ src/test/resources/env/hub/stopHub.sh
@@ -1,3 +1,9 @@
 #!/bin/sh
-kill `cat pid`
-rm pid
+
+if [ -f pid ]
+then
+    kill `cat pid`
+    rm pid
+else
+    echo "File pid doesn't exist" >> log
+fi
